Remove obsolete comments
I believe the comments originated from a time when none of the 5 virtual
methods were known, and no placeholders for them existed. This meant
later virtual functions would not bind correctly. Placeholders needed to
be added to kept the place in the virtual function table, to ensure
later functions would bind to the correct entry.